+++ tests/testthat/test_figures.R
@@ -28,6 +28,7 @@ test_that("graph_test_pq works", {
   expect_message(graph_test_pq(subset_samples(data_fungi, !is.na(data_fungi@sam_data$Time)), fact = "Time", merge_sample_by = "Tree_name"))
   expect_error(graph_test_pq(data_fungi, fact = "Height"))
   expect_error(graph_test_pq(enterotype, fact = "Enterotype"))
+  expect_error(graph_test_pq(data_fungi, fact = "tRREE_name"))
 })
 
 test_that("accu_plot works", {
@@ -35,4 +36,13 @@ test_that("accu_plot works", {
   expect_silent(accu_plot(GP_archae, "SampleType", add_nb_seq = TRUE, by.fact = TRUE))
   expect_warning(accu_plot(GP_archae, "SampleType", add_nb_seq = FALSE, by.fact = TRUE))
   expect_error(accu_plot(GP_archae))
+})
+
+
+data_fungi <- subset_samples(data_fungi, !is.na(Time))
+res_mt <- mt(data_fungi, "Time", method = "fdr", test = "f", B = 300)
+test_that("plot_mt works", {
+  expect_s3_class(res_mt, "data.frame")
+  expect_s3_class(suppressWarnings(plot_mt(res_mt)), "ggplot")
+  expect_s3_class(suppressWarnings(plot_mt(res_mt, taxa = "Genus", color_tax = "Order")), "ggplot")
 })